add: RecommendSection 컴포넌트 추가 (#19)
