+++ src/components/DetailSection.tsx
@@ -1,154 +1,13 @@
-import { Link } from "react-router-dom";
 import PainChart from "./PainChart";
 import ConditionList from "./ConditionList";
 import SectionHeader from "./PersonalReport/SectionHeader";
 import MediaSection from "./PersonalReport/Media/MediaSection";
 import FeedbackSection from "./PersonalReport/Feedback/FeedbackSection";
+import RecommendSection from "./PersonalReport/Recommend/RecommendSection";
 
 interface PropType {
   type: "media" | "feedback" | "recommend" | "pain" | "condition";
 }
-
-const personalReport = {
-  uuid: "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-  writer: {
-    id: 0,
-    name: "김파이",
-  },
-  center: {
-    id: 0,
-    name: "string",
-    code: "string",
-    phone: "string",
-    address: "string",
-    contackLink: "string",
-  },
-  member: {
-    id: 0,
-    name: "홍길동",
-    phone: "string",
-    sex: "MALE",
-    birthDate: "2023-07-19",
-    createdAt: "2023-07-19T06:05:16.520Z",
-    updatedAt: "2023-07-19T06:05:16.520Z",
-    visitedAt: "2023-07-19T06:05:16.520Z",
-  },
-  reviewRating: 0,
-  reviewContent: "string",
-  comment: {
-    hidden: true,
-    content:
-      "회원님 어깨가 빨리 개선되기 위해 제가 오늘 하신 운동이랑 저희가 제안드리는 운동 처방 같이 드렸습니다. 꼭 회원님 하루에 10회씩 2세트씩 해주세요.",
-  },
-  media: {
-    hidden: true,
-    items: [
-      {
-        type: "IMAGE",
-        uuid: "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-        url: "string",
-        thumbnailUrl:
-          "https://images.unsplash.com/photo-1575052814086-f385e2e2ad1b?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-      },
-      {
-        type: "VIDEO",
-        uuid: "3fa85f64-5717-4562-b3fc-2c963f66afa7",
-        url: "string",
-        thumbnailUrl:
-          "https://images.unsplash.com/photo-1575052814086-f385e2e2ad1b?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-      },
-      {
-        type: "VIDEO",
-        uuid: "3fa85f64-5717-4562-b3fc-2c963f66afa8",
-        url: "string",
-        thumbnailUrl:
-          "https://images.unsplash.com/photo-1575052814086-f385e2e2ad1b?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-      },
-      {
-        type: "VIDEO",
-        uuid: "3fa85f64-5717-4562-b3fc-2c963f66afa9",
-        url: "string",
-        thumbnailUrl:
-          "https://images.unsplash.com/photo-1575052814086-f385e2e2ad1b?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-      },
-      {
-        type: "VIDEO",
-        uuid: "3fa85f64-5717-4562-b3fc-2c963f66afa0",
-        url: "string",
-        thumbnailUrl:
-          "https://images.unsplash.com/photo-1575052814086-f385e2e2ad1b?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-      },
-    ],
-  },
-  archiveLink: {
-    hidden: true,
-    items: [
-      {
-        title: "string",
-        description: "string",
-        url: "https://images.unsplash.com/photo-1562771379-eafdca7a02f8?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-        site: "string",
-      },
-      {
-        title: "string",
-        description: "string",
-        url: "https://images.unsplash.com/photo-1571019613454-1cb2f99b2d8b?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-        site: "string",
-      },
-      {
-        title: "string",
-        description: "string",
-        url: "https://images.unsplash.com/photo-1575052814086-f385e2e2ad1b?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-        site: "string",
-      },
-      {
-        title: "string",
-        description: "string",
-        url: "https://images.unsplash.com/photo-1562771379-44b243dedac5?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2340&q=80",
-        site: "string",
-      },
-    ],
-  },
-  painHistory: {
-    hidden: true,
-    items: [
-      {
-        bodyCode: 0,
-        histories: [
-          {
-            date: "2023-07-19",
-            level: 0,
-          },
-        ],
-      },
-    ],
-  },
-  condition: {
-    hidden: true,
-    items: [
-      {
-        date: "2023-07-19",
-        condition: "string",
-      },
-      {
-        date: "2023-07-18",
-        condition: "string",
-      },
-      {
-        date: "2023-07-17",
-        condition: "string",
-      },
-      {
-        date: "2023-07-16",
-        condition: "string",
-      },
-      {
-        date: "2023-07-15",
-        condition: "string",
-      },
-    ],
-  },
-};
 
 const DetailSection = ({ type }: PropType) => {
   return (
@@ -157,27 +16,7 @@ const DetailSection = ({ type }: PropType) => {
       <div className="mt-2">
         {type === "media" && <MediaSection />}
         {type === "feedback" && <FeedbackSection />}
-        {type === "recommend" && (
-          <ul className="mt-2 grid grid-rows-2 grid-flow-col gap-2 overflow-hidden">
-            {personalReport.archiveLink.items.map((item) => (
-              <li key={item.url}>
-                <Link
-                  to={"#"}
-                  className="w-60 flex items-center space-x-2 border p-2 rounded-lg"
-                >
-                  <div
-                    className="w-12 h-12 rounded-md shadow-md"
-                    style={{
-                      backgroundImage: `url(${item.url})`,
-                      backgroundSize: "cover",
-                    }}
-                  />
-                  <span className="text-sm">{item.description}</span>
-                </Link>
-              </li>
-            ))}
-          </ul>
-        )}
+        {type === "recommend" && <RecommendSection />}
         {type === "pain" && <PainChart />}
         {type === "condition" && <ConditionList />}
       </div>
